Made changes related to optimization
broke down explanation canvas into BarirerwithSlit and WaveDrawing
